requirements.txt
asgiref==3.8.1
boto3==1.38.32
botocore==1.38.32
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
django-dotenv==1.4.2
django-filter==25.1
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
jmespath==1.0.1
packaging==25.0
pillow==11.2.1
psycopg==3.2.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
stripe==12.2.0
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
